Fix publishable workspace test dependency cycle
